Update osxinstall, the script now properly works.
Pweh, I managed to make it work (and applied the return statements to the applescript code, thx for that!), the only thing that does not properly work is the fact that if you cancel the installation, it won't end, if you could get a look at that, would be great. mv command keeps verbosing if the file is the same, even with the -f modifier.