Update osxinstall, prevents error if rlvm wasn't installed
It was a problem with the symlinks and relative paths ("~/temp" is not interpretable by applescript, but echo-ing that path returns the entire folder path "/Users/user/temp", fixing the problem).

Also, applescript fails if you specify a non-existing folder as the default location of the prompt of seldir() function, so added an if statement that checks if the folder exists.

The script now also tries to navigate to the usual installation folder on wine, if you specify that you have already installed Tomoyo After. If the folder does not exist, it navigates to the home directory. It makes the same if the rlvm installation is not detected when choosing the font installation path ("~/.rlvm/" not existing, in fact).